Reading state creation

+++ SK_EPReader/Classes/ChapterLoader.m
@@ -58,7 +58,7 @@
     
     NSString *insertRule1 = [NSString stringWithFormat:@"addCSSRule('html', 'padding: 0px; height: %fpx; -webkit-column-gap: 0px; -webkit-column-width: %fpx;')", webView.frame.size.height, webView.frame.size.width];
     NSString *insertRule2 = [NSString stringWithFormat:@"addCSSRule('p', 'text-align: justify;')"];
-    NSString *setTextSizeRule = [NSString stringWithFormat:@"addCSSRule('body', '-webkit-text-size-adjust: %ld%%;')", self.chapter.fontPercentSize];
+    NSString *setTextSizeRule = [NSString stringWithFormat:@"addCSSRule('body', '-webkit-text-size-adjust: %ld%%;')", (unsigned long)self.chapter.fontPercentSize];
     
     
     [webView stringByEvaluatingJavaScriptFromString:varMySheet];
